pyproject.toml
[build-system]
requires = ["setuptools>=42", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lotusrpc"
description = "A code generator for remote procedure calls on embedded systems"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "T Zijnge" }]
keywords = [
    "code generation",
    "embedded systems",
    "C++",
    "RPC",
    "remote procedure call",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: C++",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Embedded Systems",
]
dependencies = [
    "code-generation==2.3.0",
    "pyyaml==6.0.2",
    "jsonschema==4.25.0",
    "click==8.1.8",
    "typing-extensions==4.14.0",
    "colorama==0.4.6",
]
dynamic = ["version"]

[project.optional-dependencies]
serial = ["pyserial==3.5"]

[project.urls]
homepage = "https://github.com/tzijnge/LotusRpc"
"Bug Tracker" = "https://github.com/tzijnge/LotusRpc/issues"
"Documentation" = "https://tzijnge.github.io/LotusRpc/"
"Source Code" = "https://github.com/tzijnge/LotusRpc"

[project.scripts]
lrpcg = "lrpc.lrpcg:run_cli"
lrpcc = "lrpc.lrpcc:run_cli"

[tool.setuptools.package-data]
"lrpc.schema" = ["lotusrpc-schema.json"]
"lrpc.resources.cpp" = ["Server.hpp", "Service.hpp", "EtlRwExtensions.hpp"]
lotusrpc = ["py.typed"]

[tool.setuptools]
include-package-data = true

[tool.setuptools_scm]
local_scheme = "no-local-version"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]

[tool.black]
line-length = 120

[tool.pylint]
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
]
max-attributes = 12

[tool.pylint.format]
# line length is handled by black. Any remaining long lines cannot be formatted properly
disable = ["line-too-long"]
allow-reexport-from-package = true

[tool.mypy]
files = ["src", "tests"]
strict = true
# Increase strictness of checks
disallow_any_decorated = true
strict_optional = true
warn_no_return = true
warn_unreachable = true

# Display statistics after running
# dump_type_stats = true
# dump_inference_stats = true

[tool.coverage.run]
branch = true
